% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_four_trajectories.R
\docType{data}
\name{dat_four_trajectories}
\alias{dat_four_trajectories}
\title{Four trajectories}
\format{
An object of class \code{"data.frame"}
}
\usage{
data("dat_four_trajectories")
}
\description{
Example data for trajectory analysis. Totally artificial.
}
\examples{

data("dat_four_trajectories")
head(dat_four_trajectories)


}
\references{
this dataset was artificially created for the rainR package
}
\keyword{datasets}
