% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_one_trajectory.R
\docType{data}
\name{dat_one_trajectory}
\alias{dat_one_trajectory}
\title{One single trajectory}
\format{
An object of class \code{"data.frame"}
}
\usage{
data("dat_one_trajectory")
}
\description{
Example data for trajectory analysis. Totally artificial.
}
\examples{

data("dat_one_trajectory")
head(dat_one_trajectory)


}
\references{
this dataset was artificially created for the rainR package
}
\keyword{datasets}
