% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_max_acceleration.R
\name{index_max_acceleration}
\alias{index_max_acceleration}
\title{Time-point of maximum acceleration}
\usage{
index_max_acceleration(x_vector, y_vector, absolute = FALSE)
}
\arguments{
\item{x_vector}{x-coordinates of the executed trajectory.}

\item{y_vector}{y-coordinates of the executed trajectory.}

\item{absolute}{Should negative accelerations (i.e., deceleration)
be included? Defaults to \code{FALSE}.}
}
\value{
Single number indicating the index of peak acceleration.
}
\description{
Computes the index of the peak acceleration of a trajectory,
defined by vectors of x and y coordinates, and assumed to be equidistant
in time.
}
\details{
The order of the supplied vectors indicates timeadjacency.
}
\examples{
data("dat_one_trajectory")
index_max_acceleration(dat_one_trajectory$xvals, dat_one_trajectory$yvals)

}
\references{
Wirth, R., Foerster, A., Kunde, W., & Pfister, R. (2020).
Design choices: Empirical recommendations for designing two-dimensional
finger tracking experiments. Behavior Research Methods, 52, 2394 - 2416.
\doi{10.3758/s13428-020-01409-0}
}
