% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_max_dist.R
\name{index_max_dist}
\alias{index_max_dist}
\title{Time-point of maximum velocity}
\usage{
index_max_dist(x_vector, y_vector)
}
\arguments{
\item{x_vector}{Vector of the x-coordinates of the executed trajectory.}

\item{y_vector}{Vector of the y-coordinates of the executed trajectory.}
}
\value{
Single number indicating the index of peak velocity.
}
\description{
Computes the index of the peak velocity of a trajectory,
defined by arrays of x and y coordinates, and assumed to be equidistant in
time.
}
\examples{
data("dat_one_trajectory")
index_max_dist(dat_one_trajectory$xvals, dat_one_trajectory$yvals)

}
\references{
Wirth, R., Foerster, A., Kunde, W., & Pfister, R. (2020).
Design choices: Empirical recommendations for designing two-dimensional
finger tracking experiments. Behavior Research Methods, 52, 2394 - 2416.
\doi{10.3758/s13428-020-01409-0}
}
