% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_monotonic_along_ideal.R
\name{is_monotonic_along_ideal}
\alias{is_monotonic_along_ideal}
\title{Test if vector is monotonically increasing along the ideal trajectory}
\usage{
is_monotonic_along_ideal(
  x_vector,
  y_vector,
  x_start = NULL,
  y_start = NULL,
  x_end = NULL,
  y_end = NULL,
  strict = TRUE,
  warn = TRUE
)
}
\arguments{
\item{x_vector}{x-coordinates of the executed path.}

\item{y_vector}{y-coordinates of the executed path.}

\item{x_start}{x-coordinate of the start point of the ideal line.
Defaults to the first value in \code{x_vector}.}

\item{y_start}{y-coordinate of the start point of the ideal line.
Defaults to the first value in \code{y_vector}.}

\item{x_end}{x-coordinate of the end point of the ideal line.
Defaults to the last value in \code{x_vector}.}

\item{y_end}{y-coordinate of the end point of the ideal line.
Defaults to the last value in \code{y_vector}.}

\item{strict}{Must the values increase \emph{strictly}? Defaults to \code{FALSE},
indicating that a weak, not a strict definition of monotony is applied.}

\item{warn}{Will a warning be issued if the trajectory is not monotonic
(relative to the ideal line)? Defaults to \code{TRUE}.}
}
\value{
A length-one logical, indicating whether the trajectory is monotonic.
}
\description{
Checks if a trajectory, defined by vectors of x and y
coordinates, is monotonically increasing relative to an
ideal line passing through the start and end points.
}
\details{
Computes the orthogonal projection of the trajectory points onto the ideal
line and checks whether this (one-dimensional) projection is monotonic.
All objects of length 0 or 1 are monotonic.
Data with missing values will not pass the check.
}
\examples{
is_monotonic_along_ideal(c(0,0,1), c(0,1,1))
is_monotonic_along_ideal(c(0,0,0,1), c(0,1,1,1), strict=TRUE, warn=FALSE)

}
\references{
Wirth, R., Foerster, A., Kunde, W., & Pfister, R. (2020).
Design choices: Empirical recommendations for designing two-dimensional
finger tracking experiments. Behavior Research Methods, 52, 2394 - 2416.
\doi{10.3758/s13428-020-01409-0}
}
