% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wuelib-package.R
\docType{package}
\name{wuelib-package}
\alias{wuelib}
\alias{wuelib-package}
\title{wuelib: Trajectory Analyses for Cognitive Psychologists}
\description{
Bundles together several functions that are useful for analyzing mouse movement (and other 2-D trajectory) data in the context of cognitive psychology. 'wuelib' is a library that is maintained by the Cognition & Behaviour team of the University of Wuerzburg.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mc-schaaf/wuelib}
  \item Report bugs at \url{https://github.com/mc-schaaf/wuelib/issues}
}

}
\author{
\strong{Maintainer}: Moritz Schaaf \email{moritz.schaaf@uni-wuerzburg.de}

Authors:
\itemize{
  \item Robert Wirth \email{robert.wirth@uni-wuerzburg.de}
  \item Roland Pfister \email{mail@roland-pfister.net}
}

}
\keyword{internal}
