% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Area Under the Curve}
\usage{
auc(
  x_vector,
  y_vector,
  x_start = NULL,
  y_start = NULL,
  x_end = NULL,
  y_end = NULL,
  cumulative = FALSE,
  RP_style = FALSE
)
}
\arguments{
\item{x_vector}{Vector of the x-coordinates of the executed trajectory}

\item{y_vector}{Vector of the y-coordinates of the executed trajectory}

\item{x_start}{x-coordinate of the start point of the ideal trajectory}

\item{y_start}{y-coordinate of the start point of the ideal trajectory}

\item{x_end}{x-coordinate of the end point of the ideal trajectory}

\item{y_end}{y-coordinate of the end point of the ideal trajectory}

\item{cumulative}{Whether one single number (cumulative=F)
or an array of cumulative AUCs should be returned}

\item{RP_style}{Should areas that stem from a change of direction be added
or subtracted?}
}
\value{
AUC as single number or as vector of cumulative AUCs.
}
\description{
Computes the (cumulative) AUC of a point defined by x and y
coordinates, as compared to an ideal trajectory, defined by the
start and end points.
Importantly, the ideal trajectory is thought of as being of infinite length.
Works via matrix-rotation.
}
\examples{
data("dat_one_trajectory")
auc(dat_one_trajectory$xvals, dat_one_trajectory$yvals)

}
\references{
Wirth, R., Foerster, A., Kunde, W., & Pfister, R. (2020).
Design choices: Empirical recommendations for designing two-dimensional
finger tracking experiments. Behavior Research Methods, 52, 2394 - 2416.
\doi{10.3758/s13428-020-01409-0}
}
