% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{curvature}
\alias{curvature}
\title{Curvature}
\usage{
curvature(x_vector, y_vector)
}
\arguments{
\item{x_vector}{vector of the x-coordinates of the executed trajectory.}

\item{y_vector}{vector of the y-coordinates of the executed trajectory.}
}
\value{
Single number indicating the curvature.
}
\description{
Computes the curvature of a trajectory, defined by arrays
of x and y coordinates, as compared to an ideal trajectory,
as defined by the start and end points of the trajectory.
}
\examples{
data("dat_one_trajectory")
curvature(dat_one_trajectory$xvals, dat_one_trajectory$yvals)

}
\references{
Wirth, R., Foerster, A., Kunde, W., & Pfister, R. (2020).
Design choices: Empirical recommendations for designing two-dimensional
finger tracking experiments. Behavior Research Methods, 52, 2394 - 2416.
\doi{10.3758/s13428-020-01409-0}
}
