% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mousetRajectory-package.R
\docType{package}
\name{mousetRajectory-package}
\alias{mousetRajectory}
\alias{mousetRajectory-package}
\title{mousetRajectory: Mouse Trajectory Analyses for Behavioural Scientists}
\description{
Helping psychologists and other behavioural scientists to analyze mouse movement (and other 2-D trajectory) data. Bundles together several functions that compute spatial measures (e.g., maximum absolute deviation, area under the curve, sample entropy) or provide a shorthand for procedures that are frequently used (e.g., time normalization, linear interpolation, extracting initiation and movement times). For more information on these dependent measures, see Wirth et al. (2020) \doi{10.3758/s13428-020-01409-0}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mc-schaaf/mousetRajectory}
  \item \url{https://mc-schaaf.github.io/mousetRajectory/}
  \item Report bugs at \url{https://github.com/mc-schaaf/mousetRajectory/issues}
}

}
\author{
\strong{Maintainer}: Roland Pfister \email{moritz.schaaf@uni-trier.de} (\href{https://orcid.org/0000-0002-4429-1052}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Solveig Tonn \email{tonn@uni-trier.de} (\href{https://orcid.org/0000-0001-5254-8391}{ORCID})
  \item Moritz Schaaf \email{moritz.schaaf@uni-trier.de} (\href{https://orcid.org/0000-0002-9959-2928}{ORCID})
  \item Robert Wirth \email{robert.wirth@uni-wuerzburg.de} (\href{https://orcid.org/0000-0001-8446-1880}{ORCID})
}

}
\keyword{internal}
