% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_crosses.R
\name{point_crosses}
\alias{point_crosses}
\title{Number of times a point is crossed}
\usage{
point_crosses(numeric_array, relevant_point = 0)
}
\arguments{
\item{numeric_array}{Array of numbers ordered by their time of appearance.}

\item{relevant_point}{Number which has to be crossed.}
}
\value{
Number of times that \code{numeric_array} crosses
the \code{relevant_point}.
}
\description{
Checks how often a certain point is being crossed by an number
sequence where the order of the array indicates timeadjacency.
}
\examples{
data("dat_one_trajectory")
point_crosses(dat_one_trajectory$xvals)

}
\references{
Wirth, R., Foerster, A., Kunde, W., & Pfister, R. (2020).
Design choices: Empirical recommendations for designing two-dimensional
finger tracking experiments. Behavior Research Methods, 52, 2394 - 2416.
\doi{10.3758/s13428-020-01409-0}
}
