% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampen.R
\name{sampen}
\alias{sampen}
\title{SAMPle ENtropy}
\usage{
sampen(
  timeseries_array,
  dimensions = 2,
  tolerance = 0.2,
  standardize = TRUE,
  use_diff = FALSE
)
}
\arguments{
\item{timeseries_array}{Array of numbers over which the sampEn is to be
computed.}

\item{dimensions}{Number of embedding dimensions for which to compute the
sampEn. Sometimes also called "template length".}

\item{tolerance}{Tolerance for the comparisons of two number sequences.}

\item{standardize}{Whether to standardize the timeseries_array.}

\item{use_diff}{Whether to use the differences between adjacent points.}
}
\value{
Single number indicating the sample entropy for the given parameters.
}
\description{
Computes the sample Entropy, as given by
Richman & Moorman (2000) <\doi{10.1152/ajpheart.2000.278.6.H2039}>.
}
\examples{
data("dat_one_trajectory")
sampen(dat_one_trajectory$xvals)

}
\references{
Wirth, R., Foerster, A., Kunde, W., & Pfister, R. (2020).
Design choices: Empirical recommendations for designing two-dimensional
finger tracking experiments. Behavior Research Methods, 52, 2394 - 2416.
\doi{10.3758/s13428-020-01409-0}
}
